scripts/data_ingestion/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
google-cloud-bigquery==3.17.1
pydantic==2.6.1
fsspec==2025.2.0
gcsfs==2025.2.0